Schema: Simplify preferred capitalisation implementation
Because the parser regards all TM schema symbols to be case insensitive, we
have no reason to store the all-caps name separately from the capitalised
name in the symbol table. The capitalised name is still preferred for
generating the programmatic identifiers, so it's still specified in
TmSchema.h.

Basically, there was no reason to ever separate the two other than for
validation purposes, so in RELEASE builds, this extra data is skipped. It
is still contained in debug builds for the new validation function.

Alongside this commit, a validation function has been added to the schema
test code to ensure that the contents of the table make sense. In order to
implement this, code regarding schema testing had to be reworked a bit.

Using the schema validator, I caught some flaws in TmSchema.h that I fixed
up in this commit too.

Signed-off-by: Isabella Lulamoon <kawapure@gmail.com>