Improve coverage of schema testing/validation
It's always nice to ensure that we didn't make any mistakes, so I edited
the schema validation to cover more cases, such as non-sequential IDs or
illegal part/state values (in case someone accidentally does zero-indexing)

For all intents and purposes, TM_PART, TM_STATE, and TM_ENUM macros are all
the same. They are not covered by this validation, so you can currently
mix them up and confuse the readers of the source code, but the program
will still operate fine. I may add validation for this in the future, but
right now I should focus on bringing in more parts & states that were added
after XP.

Signed-off-by: Isabella Lulamoon <kawapure@gmail.com>