Work on text scanner for INI parser
This is early work, and it's a bit messy. I'm currently struggling to deal
with string length boundaries and automatic buffer sizing in a nice manner.

XP avoided these issues by having a hard cap on input (i.e. line can't be
more than 256 characters) or just didn't check at all, but XP only needed
to account for Microsoft-controlled input (at least, in theory). We have
all sorts of arbitrary user files to parse, regardless of file encoding or
comment style or whatever, and we're building a proper compiler, so I think
that we ought to be more accommodating to weird input that doesn't fit
predetermined buffer sizes.

Thus, I'm thinking we should roll a string object of some kind, or maybe
just use std::string. We've avoided a lot of C++isms in restyle, but I
think it's a good option rather than zero-terminated strings and character
count arguments, at least for a source code parser. Results in less-messy
functions and allows us to hide away buffer reallocation.

Signed-off-by: Isabella Lulamoon <kawapure@gmail.com>