Factorise IniParser arenas; move declarations to private header...
+ Replace node tree with a flat array map. We don't need a node tree even
though the format we're implementing has depth, because the depth cannot
nest arbitrarily and is constant.

+ Removed C preprocessor <> quote literals. We will simply not support
these because we aren't implementing a true C preprocessor. Inclusion will
only support other INI files, and we will not have an include directory of
library INI files to choose from, so it's a pointless thing to account for.

Signed-off-by: Isabella Lulamoon <kawapure@gmail.com>