IniParser: Move arenas to use linked list for expansion
This is better than removing the memory and reallocating, which is not
memory-safe. In that case, it is possible that the base pointer for the
arena data could change, which would result in all existing pointers being
invalidated, which is really bad.

With a linked list, the location of the arena in memory is kept constant.
The only downside is that it's partitioned after the base size, so the
memory is likely to be fragmented somewhere. This isn't a big issue, since
we only really need things to be mostly contiguous.

This also includes some changes to support static polymorphism, which is
required because we now need to create instances of any of the child
arena classes from within the base implementation.

Some miscellaneous code cleanup was also done in this commit.

Signed-off-by: Isabella Lulamoon <kawapure@gmail.com>