Restructure CTBaseArena, miscellaneous changes...
Out of all the miscellaneous changes, I think the most important one is
schemautils.cpp, which will be used to standardise and accelerate schema
lookups.

I accidentally wrote CTBaseArena with standard C++ semantics, whereby class
definition and class implementation are separate. This is kind of useless
with templates, as you cannot forward declare a template and implement its
methods in a separate header file.

In this commit, CTBaseArena was moved to a new home, arena.h, where it will
remain for any future use we desire.

Signed-off-by: Isabella Lulamoon <kawapure@gmail.com>