Fix inline functions in scanner.cpp
An easy mistake you can make in C++ is splitting the declarations of inline
functions and their implementations between two different files. This only
works properly, however, if the header file (= the file in which the
declaration is made) is in the same translation unit as the implementation
file. In other words, it only really works if you ONLY include the .h
in the corresponding .cpp file and you don't include it in any other file.

This is because, of course, inlining is a process which the compiler does,
and the compiler does not have access to other translation units. Smart
linkers (i.e. the MSVC linker) can also perform link-time inlining, so it's
not a big deal. In this case, it can happen across object files (just think
"translation units") because the linker isn't stupid and limited in its
world view like C and C++ compilers are made to be.

Another note is that this type of error is particularly evil because it can
cause a compile-time error, but only sometimes.

This is all obvious stuff when you think about it, but sometimes you just
end up writing code without thinking about it, so you make silly mistakes
like these.

Signed-off-by: Isabella Lulamoon <kawapure@gmail.com>