Big INI parser work dump (now Restyle compiles again!!)
Past 15-ish hours of work here.

This includes preprocessor defines to disable unfinished code which won't
compile (right now this pertains, ironically, to the implementation for the
preprocessor of the INI parser). This fixes Restyle compilation.

Work has begun on writing the actual parser for the INI-like format used by
Restyle projects.

Signed-off-by: Isabella Lulamoon <kawapure@gmail.com>