INI Parser: Restructure to use namespace Restyle
This cleans up the codebase of a lot of repeated Restyle:: references. We
aren't exactly clear on when to use that namespace and when not to. The
schema was originally designed to use it in order to avoid conflicting with
the top-level definitions of the same name that the Windows SDK provides in
its UxTheme headers, but since we need to access our custom Theme Manager
schema definitions quite frequently, we end up using the Restyle namespace
a lot within other files. It's probably worth just restructuring everything
to exist under the Restyle namespace.

Signed-off-by: Isabella Lulamoon <kawapure@gmail.com>