INI Parser: Optimise boolean values in CValueArena
Boolean values, by nature, do not require unique storage. It is wasteful to
allocate memory for each boolean value. It is only worth consideration for
faithfully decompiling weird code, which has a non-1 set boolean value. For
compilation, we strictly only allow two values to pass through. If we were
to allow manual override on enums and booleans, then we shall not modify
this function - instead, call a different function for allocation of
manual overrides.

Signed-off-by: Isabella Lulamoon <kawapure@gmail.com>